Add files via upload
The files include the Game design Christian and Jonathan came up with which is a Mario bros style of game. It also consists of the animated character via a gif, the background split into multiple layers as well as all the layers combined as well as the design web page, resource web page, layer web page, and the games web page.  They are not finished at this point the game web page needs a bit of layout work and the others just need some background images. The Index html needs a logo still.