Add files via upload
This is the final version of everything for the Mario Bros sliding Game. Compared to the previous version I have added a logo, as well as edited some of the layout issues. I also had to fix a few minor background issues in the different layers. There is a total of 13 separate documents most being pictures. This will be the complete final edit of everything before presenting. Most everything changed in this version was layout of the web pages.