Add files via upload
These are the updated versions of the website for the Mario Bros game. The stylesheets was newly created and contains all the styling for the pages. The Mario background picture was created by Colleen. There was a lot changed in these files most all the formatting has now been changed as well as centering and placement of text. A menu bar has been added to all the pages now so navigation is much easier. Overall the pages were cleaned up and made to look more presentable.